requirements.txt
dash
dash_auth
dash_table
pandas
python-dotenv
requests